Renamed files to match the naming convention, updated swagger types
