+++ apps/official-journal-api/src/dto/categories/journal-getcategories-query.dto.ts
