+++ apps/official-journal-api/src/dto/categories/journal-getcategories-responses.dto.ts
